Web project added

+++ Expenses.BL/Entities/Account.cs
@@ -4,11 +4,11 @@ namespace Expenses.BL.Entities
 {
     public class Account : IUnique
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
 
         public string Name { get; set; }
 
-        public int CurrencyId { get; set; }
+        public long CurrencyId { get; set; }
 
         public double Amount { get; set; }
